Fix of problems introduced in previous commit.
This is what is now completed: the meaning of names for groups of nonterminals in Grin abstract syntax are slightly changed:
AllAdapt (that is, GrAdapt and GrAdaptL) are now part of AllGrVal, and
AllSplit (that is, GrSplit and GrSplitL) are now part of AllGrPat.
This way, they don't have to be dealt with separately (and cannot be forgotten) in attributes that process names, environments etc.