The full prelude now survives all Grinc transformations. Does not generate Silly code yet, however.
Because the prelude contains many dictionaries, there are many indirect calls, which in Grin amount to the use of A tags.
Static analysis of them creates many impossible cases, which in some instances break code generations.

Innovations that were necessary to at least be able to generate low-level Grin are:
- treating GrTag_Any differently
- restoring the invariant that Apply only takes pointers, not nodes
- in the static analysis, the occur check introduced earlier took too much time in complicated cases. I removed it now, which is possible by not going into recursion at all. The function where it was used was already part of a fixpoint process, so we don't need the local recursion anyway.
- a new transformation that removes "dead" bindings, i.e. bindings of which static analysis reveals that it can never be called