Completed first runnable version of polarity inference. Run version 17 with the verbose flag to see a pretty print of the inferred polarity signatures for type constructors below the other inferred signatures of a declaration group.
Polarities are similar to kinds, with the key difference that 'star' is 
replaced by 'covariant', 'contravariant' or 'invariant' and there is a 
negation operator (corresponding to the flip of polarities for the 
arrow type constructor). Polarities are polymorhpic and there is a 
simple subtyping relation on them (covariant/contravariant weakens to 
invariant).

For example:

> ~/EHC/bin/17/ehc -v --stopat=3 ~/EHC/test/regress/17/multiple.eh

This file contains some data type definitions, and inference yield the 
following signatures:

> data Arr a b = Arrow (a -> b)
>   ::: forall d . ^d -> d -> d

> data Tuple a b = Tuple a b
>   ::: forall d . d -> d -> d

> data GGRose f a = GGBranch (f a (GGRose f a))
>   ::: forall d1 . forall d2 . (d2 -> d1 -> d1) -> d2 -> d1

And some data types to experiment with some type expressions:

> data InfArrList a = IAL (GGRose Arr a)
>   ::: forall d . ^d -> d

> data InfNestedTuple a = INT (GGRose Tuple a)
>   ::: forall d . d -> d

Other examples include mutual recursive data types, and data types 
where a type parameter occurs both positively and negatively.

There are probably a lot of bugs and mistakes, but at the moment I can 
already get the right signatures for many data types. There is at least 
one thing I have to look into, of which I don't know if it is a bug or 
a feature. Namely, or data types with conflicting 
polarities, for example:

> data Conflict a = C (a -> a)
>   ::: Invariant -> Invariant

Although the polarity for the argument is correctly inferred as 'a', 
the other part is too strict, since I want as signature:

> Conflict ::: forall delta . Invariant -> delta

I know what causes this difference. Just have to think about whether or 
not I want this and how to solve it.