Unreachable bindings removed earlier; more specialized App functions generated.
Now bindings that are removed anyway need not be cleaned, checked and numbered,
and do not waste unique numbers and thus array space in HPT analysis.
Gain: somewhat faster compilation, especially with large preludes.

For each #A/n tag in the program, an app function of arity n must be present.
This is generated in the BuildAppBindings transformation.
Formerly, a single function was generated for each arity.
Now, a specialized version is generated for each #A/n occurence.
This prevents unnecessary confusion of parameters and results of these functions in HPT analysis, thus a better approximation, thus less cases to be distinguished, thus faster code.