Fixed the last* two problems in HPT analysis. Also included some trivial example programs in the nofib suite.
When an abstract heap node's tag indicates that the node is unevaluated, the heap node can also contain the result of the unevaluated function.
This was already done for #F-nodes, but it should also be done for #A-nodes. Adapted the tagFun function to make it so.

If y is the evaluation result of x, y can have all final (#C and #P) tags that x has. If x can have #A/f tags, y can also have final tags that are the result of f.
This already worked for #C and #P tags in the result of f, but is now extended to the transitive closure for #A tags in the result of f.

* of this week