* Bugfix in Shortcut optimization (See below) * Removed todo about the above bug in my thesis.
The Shortcut transformation tries to inline variables where possible.
One assumption it makes, is that the tree is SSA, except for variables
used in cases.
But this does not hold for the global variable RP, which is clobbered
after each function call.
 
For example: we have the silly code
x := RP[1]
CALL fib( a )
y := RP[1]
CALL fib( b )
z := primAddInt( x, y )

Then Shortcut will optimize it to
CALL fib( a )
CALL fib( b )
z := primAddInt( RP[1], RP[1] )

which is obviously not correct.

Not inlining RP is possible, but a waste. So we use the following strategy:
    * Collect all variables that are an alias of an RP[x]
    * When we encounter a call, we invalidate all variables in the previous
      mentioned Set
    * If we find an Variable that that refers to an variable in the invalidated
      set, we know that the RP[x] variable is used after a new Call
    * The Set with Variables that are invalid uses of RP[x] variables is used
      to test if an variable can be inline. If it is in the Set, it cannot.