When doing full program analysis, we now let the compiler merge dictionaries for instances with the class-default definitions directly, instead of the Faxen-algorithm. Only for instances with context we fall back on the Faxen algorithm.
How it's done:
* We extended Core with an annotation mechanism that expresses the
intention of definitions of classes, instances, and default-functions.
Roughly speaking, the names of the functions defined are mentioned
in the header.
* A new Core2Core transformation (LiftDictFields) ensures that the
functions mentioned are available on toplevel.
* We extended Grin with a similar annotation mechanism. The Core2Grin
transformation translates the annotations.
* A new Grin2Grin transformation exploits the annotations for replacing
the function that generates instance-dictionaries with directly storing
the dictionary in a global variable.

What is not done:
If an instance has a context, i.e. the dictionary-generator is parameterized
by another dictionary, we can't do the transformation. Examples are:
    instance (Integral a) => Num (Ratio a)
    instance (Eq a)       => Eq  ([a])
However, we can do the more common situation of classes with a context, e.g.
    class (Eq a) => Num a

On the fly, we corrected some errors elsewhere in the pipeline:
* Opts.chs: in version 99, OptEmitBytecode is true by default.
This was not switched off when the commandline asks for full program analysis.
* Lowergrin.cag: the simplified way we update HPT tables since rev.1142
wrongly assumed that new HPT items were numbered contiguously.
(As a patch, we now fill the gaps with dummy Bottom items, but we still need
to investigate where the gaps originate from.)
* Grin/ToSilly.cag: when allocating heap records, it was assumed that thunks
would only be updated by final values. This is, however, not the case for records
that initially store a Hole-tag. Also, code generated for FetchUpdate copied
one filed too much.