Changes to shuffle to accomodate the notion of aspects, allowing configuration orthogonal to variants. The idea is to have a two dimensional configuration space, one for stepwise specification, one for aspects of the system. For now, while this feature is incorporated into ehc+shuffle, the following aspects are defined (in Makefile variabe EHC_ASPECTS): base hmty codegen. For example, this then allows the omission of codegeneration (by omitting 'codegen'), and the replacement of the type system. Shuffle now accepts the following notation:
%%[(8 codegen)
%%]

which previously only could be:

%%[8
%%]

On the commandline option --gen-reqm="(8 base hmty codegen)" specifies
that this (and other) chunk for this (and other) aspect should be
included.

The chunk offers the aspect 'codegen' and variant 8. The unpleasant
consequence of this additional feature is that all source code files
have to be modified to incorporate this extra meta information. This is
not completed yet, only done for EH/Core to try out whether it does not
break the system.

Furthermore, a start has been made with the documentation of shuffle on
a wiki page, reachable from the EHC web. Previous documentation still
exists as a pdf file but is outdated.