More libraries: Foreign related modules, etc etc. Basically preparing for dealing with IO in Haskell (so C can be avoided, which also would make full prog analysis much more difficult).
Use of gmp library has been encapsulated in the RTS, as it caused weird duplication of function definitions.
This did depend on the choice of C standard, which has now been fixed on gnu99 on all platforms, and did vary previously.
Hopefully this fixes some linux compilation problems, but checking that will be done after this commit.
It breaks one regression test, but this one is related to IO in a C primitive and exceptions; the intention is to lift this stuff to Haskell anyway, so it is left unfixed.

Warning:
- Rerun ./configure
- Module Text.Read contained some code triggering an offside parsing error in uulib.
  This has been fixed but since I do not want to push users in forcing to download uulib immediately I don't undo the workaround.