Additional modules, related to FFI
Change in heuristic for quantifier placement in type annotations. For example,

f1 :: a -> c

would be interpreted as

f1 :: forall a . a -> forall c . c

whereas the Haskell interpretation is

f1 :: forall a c . a -> c

The difference frequently led to typing errors because partial
application would yield a polymorphic type, which on some locations
impredicatively would be propagated. This would indirectly lead to more
polymorphism then necessary and also to ambiguity w.r.t. type class
predicates. This can all be fixed by providing the type signature, but
happens so often that for backwards compatibility it is more convenient
to take as the default the Haskell interpretation and leave it up to the
programmer to specify otherwise (which of course still can be done).