Merge OneBigGrin onto the trunk.
Merging of modules into one (for full program analysis) is now done after
generating Grin, instead of after generating Core. The old behaviour is still
available in variant < 100, with the option "--early-mod-merge=yes".

This commit also fixes the Grin parser and tweaks the Grin pretty printer, to
make caching of Grin possible.

Missing grin files are generated on-the-fly from core files, so it should Just
Work (tm) without rebuilding the library. If it doesn't (and/or you encounter a
"Prelude.(!!): negative index"), you should remove and rebuild the library.