Work on typed core, or better said, prerequisites: - quantified types also need their kind bound together with the quantified tyvar:   - this info must be propagated, into quantification, and out of instantiation   - because of kind polymorphism we may need to quantify over kinds for values as well
As an example, look at:

data D f a = DD (f a)

The kind of the type constructor D is (not fully implemented yet):
  D :: forall a . (a -> *) -> a -> *
The type of the value constructor DD is:
  DD :: forall* k . forall a(k -> *) . forall b . a b -> D a b

(The prettyprinting forall* is currently used for abstraction over a higher metalevel, the prettyprinting a(k->*) currently for  a :: k->*)

When used in typed core DD thus must be parameterized on three levels: kind, type, and value; D with kind and type.

The current implementation complications arise because the type inferencer has to compute the correct parameterizations to make all this fit together.