Fixed problem underlying nofib/imag/integrate benchmark test:
Cleanup of bytecode linking, replaced by a simpler and more space
efficient implementation (code size reduction of 25%). The new
implementation is a side effect of previous work on adding GC related
info, and passing it on to the GC.

In the end the mentioned 'last bug' in Jeroen's nofib selection turned
out to be in the tracing of objects. The newly copied nodes are
themselves used as the queue from which to take new objects to be
traced, but it assumes that only traceable objects are encountered.
However, the implementation breaks this assumption because because of
the consecutivity of nodes, also non traceable objects are automatically
part of this queue, and the check for 'can trace the queued object' had
to be done a second time, to prevent Double nodes to be traced. A
different queueing mechanism has been implemented as well, one which
builds the queue explicitly, in which case it is not added and the
problem cannot occur. This implementation however crashes anyway, but
may be the better solution, although theoretically slower.

Finally, there still exists the problem that unboxed values live on the
stack, and may be incorrectly traced, due to lack of information about
stack structure. The situation does not (yet) pop up in current tests,
because other checks already filter out many pointerlike values falling
outside the range managed by the GC, but surely this will happen some
day, so this problem has to be solved as well.