Work (unfinished) on infrastructure for making GC info available runtime, in the interpreter. The design chosen is to have for a stackframe a descriptor telling what may and what must not be traced for GC. There are two places where this info will/must be available: - via the BP (basepointer, markpointer), which always points to a return address, the return address always has call info, which may point to a descriptor. This describes then the part below the BP to the next BP, not above towards the SP (stackpointer). For that we need: - some instructions to also be annotated with GC info, in particular allocation, so the top part is described at instructions where this really matters.
Two alternatives were considered/pursued:
- only info about how much on the top of the stack must not be traced. This could be just enough but depends very much on which callpatterns are actually used.
- markers on the stack describing which parts must not be GC. Apart from additional runtime overhead (when not GC-ing), other values -like pointers into non GC managed memory- now occasionally are recognized as such markers. Havoc...

It is yet unclear whether this (the GC work under construction) solves the current bug under investigation. The hypothesis is that it is caused by unboxed Doubles incorrectly interpreted as pointer. The above mechanism should fix that, but pinpointing the guilty code is almost impossible. The problem occurs only for longrunning programs, and it is a somewhat random process to tweak buffer sizes and programs to obtain a short running program; something I did not succeed at for this problem. Just to give an idea, for a no tracing output generating program it might take 0.150sec to run, with tracing on it takes a couple of minutes and generates also a couple of GB trace output. The math for a 10sec program is left as an exercise.

Too small a stack also has been considered a culprit. Although increasing the stacksize makes no difference, it might be worth the trouble to rule this option out by incorporating a runtime check on remaining stacksize, which has to be done anyway.