Added check on stack overflow.
The stack overflow check pointed out a problem with stack memory leaking in tailcall variants:
- a tailapply variant of apply had to be added
- the already existing taileval had to be adapted not to consume stack memory (partially done)

The transformation of eliminating unnessary evaluations has been adapted because it helps avoiding closure building leading to the use of tailevals. It also has been adapted to take into account restrictions imposed by the HPT analysis w.r.t. change in node/pointer typing. This has not been tested yet for the C backend.

Stack overflow still occurs in the paraffins benchmark test, but not anymore due to faulty tailcall variants. It seems to be a nested sequence of closures evaluated, leading to nested calls, leading to stack overflow. Most likely this is caused by a function which has specified to enforce evaluation (by `seq` or $!), but this is not done correctly. This requires further investigation.

The effect of the above changes on performance is somewhat mixed. The additional overhead stacksize check evidently costs time (in the 3-5% range), but the stack not increasing that much means that the garbage collector has to trace less. Some benchmark test times increase, others decrease.