Made UHC compileable with GHC 7.2. That was a time consuming process but somebody had to do it...
Various source files were importing
hierarchical modules (System.IO, Data.Maybe, Data.Char, etc.) and
depricated non-hierarchical modules (IO, Maybe, Char, etc.) inconsistently.
I've replaced each the non-hierarchical import with a sequence of
hierarchical imports.

As a consequence, the dependency on the haskell98 package is not
needed anymore. Rerun autoconf, configure and make clean to ensure
that the generated cabal files do not include this dependency
anymore. This will force you to use the hierarchical modules.
Also, the sources can now be build with GHC 7.2 for which at the
moment the haskell98 package is unavailable.